+++ testcrate/tests/dag_tests/macros.rs
@@ -1,6 +1,6 @@
 #![allow(clippy::let_unit_value)]
 
-use awint::dag::{assert, assert_eq, *};
+use awint::{awint_dag::mimick, dag};
 
 use crate::dag_tests::Epoch;
 
@@ -8,6 +8,8 @@ use crate::dag_tests::Epoch;
 // currently assumes all operations succeed)
 #[test]
 fn dag_macros() {
+    use dag::*;
+    use mimick::{assert, assert_eq};
     let epoch0 = Epoch::new();
     // both trailing comma and semicolon
     let _ = inlawi!(0u1,;);